Add maven project config
This replaces the Eclipse IDE files. The project can now be checked out as a maven project. The Eclipse files referenced several .jar files in a folder called lib that weren't committed to SCM. After converting to a pom I found these dependencies weren't required to run or compile so I haven't added them to the pom.